test2

+++ lib/myLib.php
@@ -1,19 +1,25 @@
 <?php
 if(count($_FILES) > 0) {
+    $errorStr = '';
+    $types = array('image/gif', 'image/png', 'image/jpeg');
+    if (!in_array($_FILES['Books']['type']['imageFile'], $types))
+        $errorStr = 'error: неверный тип файла ' . $_FILES['Books']['name']['imageFile'] . ' *** ';
+    if ($_FILES['Books']['size']['imageFile'] > 3500000)
+        $errorStr =$errorStr . 'error: файл ' . $_FILES['Books']['name']['imageFile'] .
+            ' недопустимого размера - ' . $_FILES['Books']['size']['imageFile'] . ' (max 3500000)';
+    if (strlen($errorStr) > 0) {
+        echo $errorStr;
+        exit;
+    }
+
     $uploaddir = '..\preview\new\\';
     if (file_exists($uploaddir)){
         foreach (glob($uploaddir. '*') as $file) unlink($file);
     }
- //   $dstName = $uploaddir.$_FILES['Books']['newPreview']['name'];
-   // if (move_uploaded_file($_FILES['Books']['newPreview']['tmp_name'], $dstName)) {
     $dstName = $uploaddir.$_FILES['Books']['name']['imageFile'];
     if (move_uploaded_file($_FILES['Books']['tmp_name']['imageFile'], $dstName)) {
-        //copy($uploaddir.$_FILES['newPreview']['name'], $uploaddir . $dstName);
-        //unlink($uploaddir.$_FILES['newPreview']['name']);
-      //  echo ($uploaddir . $dstName);
         echo ($dstName);
     } else {
-        //echo ('ERRORS UPLOAD FILE');
         echo var_dump($_FILES);
     }
     exit;
